loco_pico: Introduce the code that runs on the Pi Pico 2 W
This is meant for running on the Pi Pico 2 W board that is embedded in
every loco.

Mainly, the purpose of this code is to connect to the Wifi AP exposed by
the loco_controller running on the Raspberry Pi Zero 2 W, and then send
a CONNECT operation to advertise which loco is associated with this
connection.

Once the connection established, the loco_pico waits for incoming
messages from the loco_controller (CONTROL_LOCO and LOCO_STATUS for
now).

Signed-off-by: Dede Dodu <sebnat94@hotmail.com>