sensors_pico: Adding new crate
This new crate is meant to run on a Pi Pico 2 W. The goal is to retrieve
the information from all the RFID readers and send updates about the
position of a loco to the loco_controller.

As a first step, we introduce a `detect_tag_uid` binary that can be used
as a test code for validating the reader behaves as expected.

Signed-off-by: Dede Dodu <sebnat94@hotmail.com>