sensors_pico: Implement actual sensors_pico code
Based off the experiment from detect_tag_uid program, and relying on the
existing skeleton from loco_pico, we build the sensors_pico code to make
it run on a dedicated Pi Pico 2W.

This program is in charge of gathering the information from all RFID
readers it's attached to. For every reader that identifies a loco
passing by, a message is sent to the loco controller to inform about the
loco's position.

Signed-off-by: Dede Dodu <sebnat94@hotmail.com>