loco_controller: Finer mutex granularity for the backend
This will prevent a very high level lock to be acquired, blocking all
threads while being held. Instead, we create a static hash map, since we
know all the locos that we can expect to connect, which allows only the
LocoInfo associated with each loco to be placed behind a mutex. This
allows multiple operation on different locos to be performed
simultaneously.

Signed-off-by: Dede Dodu <sebnat94@hotmail.com>