loco_controller: Handle updates from sensors
Introducing a new thread that is spawned to handle locos' location
updates from the sensors. We're expecting quite regularly to receive
some messages from the sensors, with the message containing information
about where a loco might be located.

This location info is stored as part of the LocoInfo structure
associated with each LocoId, and it is now displayed through the
loco_status HTTP API.

Signed-off-by: Dede Dodu <sebnat94@hotmail.com>