common_pico: Disable power management
By disabling power management, we don't try to save some power, which
means the board will consume the most power in this mode. The reason for
using this mode is so that we don't prevent the WiFi chip from consuming
the power it needs, which should help with maintaining a stable
connection with the loco_controller.

Signed-off-by: Dede Dodu <sebnat94@hotmail.com>