sensors_pico: Fix incorrect SensorsData array size
The generic code that was using a vector expects at most 8 readers to be
registered, and the array of CS pins is 8 in size, but the mutex array
for sharing information between the 2 threads was still made of two
entries.

Updating the mutex array so that we can share info from 8 distincts
readers.

Signed-off-by: Dede Dodu <sebnat94@hotmail.com>