actuators_pico: Change switch rails driving method
A switch rails is driven by two electric coils (one for each position),
and for that reason, we must not continuously power one position or the
other, otherwise we might end up damaging one of the coils.

Based on this observation, we've updated the wiring and the associated
logic for driving the switch rails from the relays.

This led to using 2 relays per switch rail, one relay driving one
position (Direct or Diverted).

To handle the switch 0 -> 1 and then 1 -> 0 after 500ms, we are now
spawning one task per switch rail, which is responsible for performing
the actual switch without impacting the main thread. This prevents from
slowing down the main thread when other requests are coming from the
loco_controller.

Signed-off-by: Dede Dodu <sebnat94@hotmail.com>