actuators_pico: Spawn switch_rail_controller on second core
Based on Pico's multicore ability, we've decided to run all tasks
controlling each switch rails on core1 (second core). These tasks are
not consuming much CPU time, which is why they can all be located on the
same core, while we leave the main thread running on core0 in order to
handle communication with loco_controller.

Signed-off-by: Dede Dodu <sebnat94@hotmail.com>